requirements.txt
allure-pytest
jsonschema~=4.23.0
pytest~=8.3.3
python-dotenv~=1.0.1
requests~=2.32.3
selene~=2.0.0rc9
selenium~=4.27.1
